+++ pubhubs_hub/modules/pseudonyms.py
@@ -36,6 +36,20 @@ class Pseudonym:
             on_profile_update = self.change_displayname,
         )
 
+        # Check whether ConfigChecker module is loaded.
+        # We do this here, because this module is so old it's surely included in any hub's configuration.
+        try:
+            for module_details in api._hs.config.modules.loaded_modules:
+                (module_class, module_config) = module_details
+                if module_class.__name__ == "ConfigChecker":
+                    break
+            else:
+                raise ConfigError("Cannot find ConfigChecker module; please add  '- module: conf.modules.config_checker.ConfigChecker' to the 'modules:' list in homeserver.yaml")
+        except ConfigError:
+            raise
+        except Exception as e:
+            logger.error(f"failed to check the presence of ConfigChecker: {e}")
+
     #
     # Make sure displayname is pseudonym at registration
     #