+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 *Please add a brief description of any changes to be tested and any migrations to be performed here.*
 
+- Several changes to Synapse `homeserver.yaml` configuration, including a `ConfigChecker` module that
+  will crash synapse to tell you what changes to make.  Please check login, registration, and anything
+  related to secure rooms carefully.
+
 ## 28 Feb 2024
 - New PHC session cookies to solve a CRSF, see #514.
 
@@ -10,6 +14,7 @@
   ```
   hotfixes:
     no_secure_cookies: true # browsers should allow secure cookies on insecure localhost, but they don't
+  ```
 
 ## 1 Feb 2024
 - Styling improvements: #174, #316, #370, #385